download data script

+++ scene_graph/download_data.sh
@@ -1,27 +1,44 @@
-#!/bin/bash
-mkdir -p data
-cd data
+# Execute from snorkel-tutorials/scene_graph
+# Download data
 
-# download and unzip metadata and annotations.
-wget https://www.dropbox.com/s/bnfhm6kt9xumik8/vrd.zip
-unzip vrd.zip
+ANNOTATIONS_URL="https://www.dropbox.com/s/bnfhm6kt9xumik8/vrd.zip"
+IMAGES_URL="http://imagenet.stanford.edu/internal/jcjohns/scene_graphs/sg_dataset.zip"
+SAMPLE_IMAGES_URL="https://github.com/Prof-Lu-Cewu/Visual-Relationship-Detection/trunk/samples"
+GLOVE_URL="http://nlp.stanford.edu/data/wordvecs/glove.6B.zip"
 
-# Delete the zip files.
-rm vrd.zip
-cd VRD
+if [ ! -d "data" ]; then
+    mkdir -p data
+    cd data
 
-# Download and unzip images.
-wget http://imagenet.stanford.edu/internal/jcjohns/scene_graphs/sg_dataset.zip
-unzip sg_dataset.zip
-rm sg_dataset.zip
+    # download and unzip metadata and annotations
+    wget $ANNOTATIONS_URL
+    unzip vrd.zip
 
-cd ../..
+    # Delete the zip files.
+    rm vrd.zip
+    cd VRD
 
-mkdir -p data/glove
-cd data/glove
+    if [ "$TRAVIS" = "true" ]; then
+        # Download and unzip sample images
+        mkdir sg_dataset
+        cd sg_dataset
+        svn checkout $SAMPLE_IMAGES_URL
+        cd ../..
+    else
+        # Download and unzip all images
+        wget $IMAGES_URL
+        unzip sg_dataset.zip
+        rm sg_dataset.zip
+        cd ../..
+    fi
 
-wget http://nlp.stanford.edu/data/wordvecs/glove.6B.zip
-unzip glove.6B.zip
+    mkdir -p data/glove
+    cd data/glove
 
-# Delete the zip files.
-rm  glove.6B.zip
+    wget $GLOVE_URL
+    unzip glove.6B.zip
+
+    # Delete the zip files
+    rm  glove.6B.zip
+fi
+
